+++ cms-20240330/alibabacloud_cms20240330/models/_get_service_task_request.py
@@ -9,6 +9,14 @@ class GetServiceTaskRequest(DaraModel):
         self,
         type: str = None,
     ):
+        # The task type. This parameter is required. Valid values:
+        # 
+        # - heapdump: heap dump.
+        # - LiveDebug Probe: live_debug_log_probe, live_debug_snapshot_probe, live_debug_metric_probe, live_debug_span_probe, live_debug_span_tag_probe.
+        # - LiveDebug Command: live_debug_inspect_object, live_debug_search_type, live_debug_search_method, live_debug_decompile, live_debug_get_thread_info, live_debug_get_runtime_info, live_debug_get_memory_info, live_debug_evaluate_expression, live_debug_modify_logger_level.
+        # - LiveDebug code hot replacement: live_debug_code_replace.
+        # 
+        # The value must be the same as the type specified during task creation.
         self.type = type
 
     def validate(self):