+++ cms-20240330/alibabacloud_cms20240330/models/_list_tag_resources_response_body.py
@@ -15,13 +15,13 @@ class ListTagResourcesResponseBody(DaraModel):
         request_id: str = None,
         tag_resources: List[main_models.ListTagResourcesResponseBodyTagResources] = None,
     ):
-        # The maximum number of results to return. The maximum value is 200.
+        # The maximum number of results to return. Maximum value: 200.
         self.max_results = max_results
-        # The token that is used to start the next query.
+        # The token for the next query.
         self.next_token = next_token
         # The request ID.
         self.request_id = request_id
-        # A list of tags.
+        # The label list.
         self.tag_resources = tag_resources
 
     def validate(self):
@@ -82,9 +82,9 @@ class ListTagResourcesResponseBodyTagResources(DaraModel):
         self.resource_id = resource_id
         # The resource type.
         self.resource_type = resource_type
-        # The tag key.
+        # The label key.
         self.tag_key = tag_key
-        # The tag value.
+        # The label value.
         self.tag_value = tag_value
 
     def validate(self):