+++ cms-20240330/alibabacloud_cms20240330/models/_untag_resources_shrink_request.py
@@ -12,12 +12,12 @@ class UntagResourcesShrinkRequest(DaraModel):
         resource_type: str = None,
         tag_key_shrink: str = None,
     ):
-        # Specifies whether to remove all tags from the specified resources. Valid values:
+        # Specifies whether to unbind all labels from the specified resource. Valid values:
         # 
-        # false (default): No
-        # true: Yes
+        # - false (default): No.
+        # - true: Yes.
         self.all = all
-        # A list of resource IDs.
+        # The list of resource IDs.
         # 
         # This parameter is required.
         self.resource_id_shrink = resource_id_shrink
@@ -25,9 +25,9 @@ class UntagResourcesShrinkRequest(DaraModel):
         # 
         # This parameter is required.
         self.resource_type = resource_type
-        # The tag key.
+        # The label key.
         # 
-        # You can detach tags with up to 20 tag keys.
+        # You can unbind up to 20 label keys at a time.
         self.tag_key_shrink = tag_key_shrink
 
     def validate(self):